Address PR review comments for player-night-summary
- Extract PlayerNightSummary component with PlayerNightSummaryProps
- Move getActionText to display.ts; absorb nightSummary/targetWasKilled logic
- Add getPlayer helper to lib/player-utils.ts
- Extract GameSerializationService with public testable methods
- Rebuild serialization tests in GameSerializationService.test.ts
- Remove now-redundant private methods from FirebaseGameService
- Clean up unused imports in FirebaseGameService
