+++ app/src/services/GameSerializationService.ts
@@ -1,0 +1,180 @@
+import { GameStatus } from "@/lib/types";
+import type { Game, RoleDefinition, Team } from "@/lib/types";
+import type { PlayerGameState } from "@/server/types";
+import {
+  isTeamNightAction,
+  getTeamPhaseKey,
+  getTeamPlayerIds,
+  WerewolfPhase,
+} from "@/lib/game-modes/werewolf";
+import type {
+  AnyNightAction,
+  TargetCategory,
+  WerewolfRoleDefinition,
+  WerewolfTurnState,
+} from "@/lib/game-modes/werewolf";
+import { GAME_MODES } from "@/lib/game-modes";
+
+/**
+ * Stateless serialization/sanitization helpers used by FirebaseGameService
+ * to build per-player game state. Exposed as public methods so they can be
+ * tested directly without constructing a full Firebase game service.
+ */
+export class GameSerializationService {
+  /**
+   * Extracts nightActions from the current turnState, if present.
+   * Returns undefined when the game is not playing or has no recorded actions.
+   */
+  extractNightActions(game: Game): Record<string, AnyNightAction> | undefined {
+    if (game.status.type !== GameStatus.Playing) return undefined;
+    const ts = game.status.turnState as WerewolfTurnState | undefined;
+    if (!ts) return undefined;
+    const { nightActions } = ts.phase;
+    return Object.keys(nightActions).length > 0 ? nightActions : undefined;
+  }
+
+  /** Extracts deadPlayerIds from the Werewolf turn state. */
+  extractDeadPlayerIds(game: Game): string[] {
+    if (game.status.type !== GameStatus.Playing) return [];
+    const ts = game.status.turnState as WerewolfTurnState | undefined;
+    return ts?.deadPlayerIds ?? [];
+  }
+
+  /**
+   * Extracts the night targeting state for a non-owner player.
+   * For solo roles: returns myNightTarget/myNightTargetConfirmed.
+   * For team phases: also returns teamVotes, suggestedTargetId, allAgreed.
+   */
+  extractPlayerNightState(
+    nightActions: Record<string, AnyNightAction>,
+    game: Game,
+    callerId: string,
+    myRole: RoleDefinition,
+    deadPlayerIds: string[],
+  ): Partial<PlayerGameState> {
+    const roleDef = GAME_MODES[game.gameMode].roles[myRole.id] as
+      | { teamTargeting?: boolean; team?: string }
+      | undefined;
+
+    if (roleDef?.teamTargeting && roleDef.team) {
+      const phaseKey = getTeamPhaseKey(roleDef.team as Team);
+      const action = nightActions[phaseKey];
+      if (!action || !isTeamNightAction(action)) {
+        return { myNightTarget: undefined, myNightTargetConfirmed: false };
+      }
+
+      const myVote = action.votes.find((v) => v.playerId === callerId);
+      const playerById = new Map(game.players.map((p) => [p.id, p]));
+
+      const aliveTeamIds = getTeamPlayerIds(
+        game.roleAssignments,
+        roleDef.team as Team,
+        deadPlayerIds,
+      );
+
+      const teamVotes = action.votes
+        .filter((v) => aliveTeamIds.includes(v.playerId))
+        .map((v) => ({
+          playerName: playerById.get(v.playerId)?.name ?? "Unknown",
+          targetPlayerId: v.targetPlayerId,
+        }));
+
+      const aliveVotes = action.votes.filter((v) =>
+        aliveTeamIds.includes(v.playerId),
+      );
+      const uniqueTargets = new Set(aliveVotes.map((v) => v.targetPlayerId));
+      const allAgreed =
+        aliveVotes.length === aliveTeamIds.length && uniqueTargets.size === 1;
+
+      return {
+        myNightTarget: myVote?.targetPlayerId,
+        myNightTargetConfirmed: action.confirmed ?? false,
+        teamVotes,
+        suggestedTargetId: action.suggestedTargetId,
+        allAgreed,
+      };
+    }
+
+    const myAction = nightActions[myRole.id];
+    if (!myAction || isTeamNightAction(myAction)) {
+      return { myNightTarget: undefined, myNightTargetConfirmed: false };
+    }
+    return {
+      myNightTarget: myAction.targetPlayerId,
+      myNightTargetConfirmed: myAction.confirmed ?? false,
+    };
+  }
+
+  /**
+   * Extracts sanitized night outcomes and the player's own last action for
+   * display at the start of the day. Only populated during daytime phases.
+   *
+   * nightSummary: only events where something happened (died === true),
+   * with attacker/protector info stripped — players cannot infer who acted.
+   *
+   * myLastNightAction: the target the player chose, even if their action was
+   * negated, so they can confirm their input was recorded.
+   */
+  extractDaytimeNightState(
+    game: Game,
+    callerId: string,
+    myRole: RoleDefinition,
+  ): Partial<PlayerGameState> {
+    if (game.status.type !== GameStatus.Playing) return {};
+    const ts = game.status.turnState as WerewolfTurnState | undefined;
+    if (ts?.phase.type !== WerewolfPhase.Daytime) return {};
+    const phase = ts.phase;
+
+    const nightSummary = (phase.nightResolution ?? [])
+      .filter((e) => e.died)
+      .map((e) => ({ targetPlayerId: e.targetPlayerId, died: e.died }));
+
+    const myLastNightAction = this.extractMyLastNightTarget(
+      phase.nightActions,
+      game,
+      callerId,
+      myRole,
+    );
+
+    return {
+      ...(nightSummary.length > 0 ? { nightSummary } : {}),
+      ...(myLastNightAction ? { myLastNightAction } : {}),
+    };
+  }
+
+  /**
+   * Returns the target the player chose during the preceding night, or
+   * undefined if they took no action.
+   */
+  extractMyLastNightTarget(
+    nightActions: Record<string, AnyNightAction>,
+    game: Game,
+    callerId: string,
+    myRole: RoleDefinition,
+  ): { targetPlayerId: string; category: TargetCategory } | undefined {
+    const roleDef = GAME_MODES[game.gameMode].roles[myRole.id] as
+      | WerewolfRoleDefinition
+      | undefined;
+    if (!roleDef) return undefined;
+
+    const { targetCategory: category } = roleDef;
+
+    if (roleDef.teamTargeting) {
+      const phaseKey = getTeamPhaseKey(roleDef.team);
+      const action = nightActions[phaseKey];
+      if (!action || !isTeamNightAction(action)) return undefined;
+      const myVote = action.votes.find((v) => v.playerId === callerId);
+      return myVote
+        ? { targetPlayerId: myVote.targetPlayerId, category }
+        : undefined;
+    }
+
+    const myAction = nightActions[myRole.id];
+    if (!myAction || isTeamNightAction(myAction)) return undefined;
+    return myAction.targetPlayerId
+      ? { targetPlayerId: myAction.targetPlayerId, category }
+      : undefined;
+  }
+}
+
+export const gameSerializationService = new GameSerializationService();
